+++ depdetective/scanners/__init__.py
@@ -1,9 +1,11 @@
 from depdetective.scanners.base import BaseScanner
+from depdetective.scanners.dotnet_nuget import DotnetNugetScanner
 from depdetective.scanners.node_package_json import NodePackageScanner
 from depdetective.scanners.python_pyproject import PythonPyprojectScanner
 from depdetective.scanners.python_requirements import PythonRequirementsScanner
 
 SCANNER_PLUGINS: list[type[BaseScanner]] = [
+    DotnetNugetScanner,
     PythonRequirementsScanner,
     PythonPyprojectScanner,
     NodePackageScanner,